refactor(control): split nfqueue cell state
