+++ crates/honk-core/src/control/nfqueue/mod.rs
@@ -19,6 +19,14 @@ use super::udp_endpoint::{EndpointReservation, OwnedEnqueueError, UdpEndpointPoo
 use crate::ebpf::{EbpfBackend, UdpDecisionCommitResult, UdpDecisionTransition};
 use crate::stats::StatsManager;
 
+mod cell;
+pub(super) use cell::PendingUdpIdentity;
+#[cfg(test)]
+use cell::TestVerdict;
+use cell::{
+    AdmissionGate, CellState, CleanupRequest, DropOutcome, FlowCell, FlowKey, HeldVerdict,
+    RetainedState, terminal_cell_is_stale,
+};
 pub(super) const TERMINAL_GRACE: Duration = Duration::from_millis(500);
 pub(super) const WATCHDOG_INTERVAL: Duration = Duration::from_millis(100);
 pub(super) const HARD_HOLD_TIMEOUT: Duration = Duration::from_secs(3);
@@ -60,60 +68,6 @@ pub(super) enum PendingUdpDecisionError {
     Fatal(#[from] PendingUdpFatal),
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-struct FlowKey {
-    client: SocketAddr,
-    destination: SocketAddr,
-}
-
-impl FlowKey {
-    const fn new(client: SocketAddr, destination: SocketAddr) -> Self {
-        Self {
-            client,
-            destination,
-        }
-    }
-
-    fn tuples(self) -> TuplesKey {
-        build_tuples_key(
-            self.destination.ip(),
-            self.destination.port(),
-            self.client.ip(),
-            self.client.port(),
-            IPPROTO_UDP,
-        )
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(super) struct PendingUdpIdentity {
-    key: FlowKey,
-    decision_token: u32,
-    endpoint_generation: u64,
-}
-
-impl PendingUdpIdentity {
-    fn new(key: FlowKey, decision_token: u32, endpoint_generation: u64) -> Self {
-        Self {
-            key,
-            decision_token,
-            endpoint_generation,
-        }
-    }
-
-    pub(super) const fn client(self) -> SocketAddr {
-        self.key.client
-    }
-
-    pub(super) const fn destination(self) -> SocketAddr {
-        self.key.destination
-    }
-
-    fn tuples(self) -> TuplesKey {
-        self.key.tuples()
-    }
-}
-
 // The listener consumes this result immediately; keeping the sole lease inline
 // avoids a second allocation on every staged UDP flow.
 #[allow(clippy::large_enum_variant)]
@@ -124,268 +78,6 @@ pub(super) enum NfqueueIngest {
     },
     Queued,
     Dropped,
-}
-
-#[cfg(test)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TestVerdict {
-    Accept { id: u64, mark: u32 },
-    Drop { id: u64 },
-}
-
-enum StoredVerdictGuard {
-    Kernel(VerdictGuard),
-    #[cfg(test)]
-    Test {
-        id: u64,
-        sink: Arc<Mutex<Vec<TestVerdict>>>,
-    },
-}
-
-impl StoredVerdictGuard {
-    fn accept(&mut self, mark: u32) -> Result<(), String> {
-        match self {
-            Self::Kernel(guard) => guard.accept(mark).map_err(|error| error.to_string()),
-            #[cfg(test)]
-            Self::Test { id, sink } => {
-                sink.lock().push(TestVerdict::Accept { id: *id, mark });
-                Ok(())
-            }
-        }
-    }
-
-    fn drop_packet(&mut self) -> Result<(), String> {
-        match self {
-            Self::Kernel(guard) => guard.drop_packet().map_err(|error| error.to_string()),
-            #[cfg(test)]
-            Self::Test { id, sink } => {
-                sink.lock().push(TestVerdict::Drop { id: *id });
-                Ok(())
-            }
-        }
-    }
-}
-
-struct HeldVerdict {
-    guard: StoredVerdictGuard,
-    received_at: Instant,
-}
-
-impl HeldVerdict {
-    fn kernel(guard: VerdictGuard, received_at: Instant) -> Self {
-        Self {
-            guard: StoredVerdictGuard::Kernel(guard),
-            received_at,
-        }
-    }
-
-    #[cfg(test)]
-    fn test(id: u64, received_at: Instant, sink: Arc<Mutex<Vec<TestVerdict>>>) -> Self {
-        Self {
-            guard: StoredVerdictGuard::Test { id, sink },
-            received_at,
-        }
-    }
-}
-
-enum CellState {
-    Pending {
-        started_at: Instant,
-        armed: bool,
-        cancelling: bool,
-        verdicts: VecDeque<HeldVerdict>,
-    },
-    ActiveDirect {
-        expires_at: Instant,
-        final_mark: u32,
-    },
-    Proxy {
-        expires_at: Instant,
-    },
-    Block {
-        expires_at: Instant,
-    },
-    Dead {
-        expires_at: Instant,
-    },
-}
-
-impl CellState {
-    fn terminal_expiry(&self) -> Option<Instant> {
-        match self {
-            Self::Pending { .. } => None,
-            Self::ActiveDirect { expires_at, .. }
-            | Self::Proxy { expires_at }
-            | Self::Block { expires_at }
-            | Self::Dead { expires_at } => Some(*expires_at),
-        }
-    }
-}
-
-struct FlowCell {
-    identity: PendingUdpIdentity,
-    _flow_slot: OwnedSemaphorePermit,
-    state: Mutex<CellState>,
-    changed: Notify,
-}
-
-impl FlowCell {
-    fn pending(
-        identity: PendingUdpIdentity,
-        started_at: Instant,
-        verdict: HeldVerdict,
-        flow_slot: OwnedSemaphorePermit,
-    ) -> Self {
-        let mut verdicts = VecDeque::with_capacity(1);
-        verdicts.push_back(verdict);
-        Self {
-            _flow_slot: flow_slot,
-            identity,
-            state: Mutex::new(CellState::Pending {
-                started_at,
-                armed: false,
-                cancelling: false,
-                verdicts,
-            }),
-            changed: Notify::new(),
-        }
-    }
-
-    fn terminal(
-        identity: PendingUdpIdentity,
-        state: CellState,
-        flow_slot: OwnedSemaphorePermit,
-    ) -> Self {
-        Self {
-            _flow_slot: flow_slot,
-            identity,
-            state: Mutex::new(state),
-            changed: Notify::new(),
-        }
-    }
-}
-
-fn terminal_cell_is_stale(cell: &FlowCell, decision_token: u32, now: Instant) -> bool {
-    cell.state
-        .lock()
-        .terminal_expiry()
-        .is_some_and(|expiry| expiry <= now || cell.identity.decision_token != decision_token)
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum CleanupRequest {
-    Flow(PendingUdpIdentity),
-    Token { key: FlowKey, decision_token: u32 },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RetainedState {
-    Pending,
-    ActiveDirect(u32),
-    Proxy,
-    Block,
-    DirectArmed,
-    Reject,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum DropOutcome {
-    Proxy,
-    Block,
-    Cancel,
-    Other,
-}
-
-#[derive(Debug)]
-struct AdmissionState {
-    open: bool,
-    epoch: u64,
-    in_flight: usize,
-}
-
-#[derive(Debug)]
-struct AdmissionGate {
-    state: Mutex<AdmissionState>,
-    quiesced: Notify,
-}
-
-impl AdmissionGate {
-    fn new() -> Self {
-        Self {
-            state: Mutex::new(AdmissionState {
-                open: false,
-                epoch: 0,
-                in_flight: 0,
-            }),
-            quiesced: Notify::new(),
-        }
-    }
-
-    fn open(&self) {
-        let mut state = self.state.lock();
-        assert_eq!(
-            state.in_flight, 0,
-            "NFQUEUE admission reopened before quiescence"
-        );
-        state.open = true;
-    }
-
-    fn try_enter(&self) -> Option<AdmissionTicket<'_>> {
-        let mut state = self.state.lock();
-        if !state.open {
-            return None;
-        }
-        state.in_flight = state
-            .in_flight
-            .checked_add(1)
-            .expect("NFQUEUE admission counter overflow");
-        Some(AdmissionTicket {
-            gate: self,
-            epoch: state.epoch,
-        })
-    }
-
-    async fn close_and_wait(&self) {
-        {
-            let mut state = self.state.lock();
-            if state.open {
-                state.open = false;
-                state.epoch = state
-                    .epoch
-                    .checked_add(1)
-                    .expect("NFQUEUE admission epoch overflow");
-            }
-        }
-        loop {
-            let notified = self.quiesced.notified();
-            if self.state.lock().in_flight == 0 {
-                return;
-            }
-            notified.await;
-        }
-    }
-}
-
-struct AdmissionTicket<'a> {
-    gate: &'a AdmissionGate,
-    epoch: u64,
-}
-
-impl Drop for AdmissionTicket<'_> {
-    fn drop(&mut self) {
-        let quiesced = {
-            let mut state = self.gate.state.lock();
-            debug_assert!(state.epoch == self.epoch || !state.open);
-            state.in_flight = state
-                .in_flight
-                .checked_sub(1)
-                .expect("NFQUEUE admission ticket underflow");
-            state.in_flight == 0
-        };
-        if quiesced {
-            self.gate.quiesced.notify_waiters();
-        }
-    }
 }
 
 pub(super) struct PendingUdpVerdicts {
